+++ app/src/main/kotlin/org/monsing/auth/AuthService.kt
@@ -2,7 +2,7 @@ package org.monsing.auth
 
 import org.monsing.auth.jwt.TokenManager
 import org.monsing.auth.jwt.TokenPayload
-import org.monsing.auth.oauth.OauthAdaptor
+import org.monsing.auth.oauthhandler.OauthAdaptor
 import org.monsing.member.GenderType
 import org.monsing.member.Member
 import org.monsing.member.MemberRepository
